Pass exceptions to receiver when communicating
When evaluation on a non-main machine caused an exception, the
thread would terminate without communicating anything, and the
error message would not appear anywhere (if on a non-main machine),
leaving the system in an blocked state waiting for the data which
caused the exception.
This patch fixes the issue by passing exceptions to the receiver
instead of the data if an exception occurs. Location information
is added to the exception (once), and if the respective data were
needed for the main computation, the exception will (eventually)
propagate to the main process on the main machine and be displayed,
terminating the program.
Exceptions which occur in unnecessary evaluations will still remain
invisible. Changes are isolated inside the write function in Trans.