fixed HC flag name
The flag to give a GHC to use is --with-hc, not --with-compiler,
so we only check for that one when building edenmodules. This is
according to the latest cabal user guide: The value of flag
--with-compiler is passed to the configure script as --with-hc, see
http://www.haskell.org/cabal/users-guide/installing-packages.html#setup-configure

The GHC build system, however, passes "--with-compiler=ghc" to
the configure script (and always "ghc"). Therefore, this does _not_
fix the build problem that an Eden-GHC with in-tree edenmodules must
use build type "Simple".