Add a Trans a to the Comm type, to justify Trans constraints
With GHC 7.11, a few crucial type signature contexts (Trans _) were
considered redundant and warned about. It turns out that the root cause
is the Communicator type (also known as ChanName) not indicating its
internal communication. A (rank-2) constraint Trans a to the argument
of Comm now propagates to the Eden constructs and justifes their
(certainly non-redundant) constraints.